add single to bot
