+++ modules/bot/controller.js
@@ -2,16 +2,12 @@ const service = require("./service");
 const controllers = require("./controllers");
 
 class ThreadController {
-  static async addSingleThread(req, res) {
-    controllers.addSingleThread(req, res, service);
+  static async gitHubPost(req, res) {
+    controllers.gitHubPost(req, res, service);
   }
-  static async addMultipleThread(req, res) {
-    controllers.addMultipleThread(req, res, service);
-  }
-  static async getAccessToken(req, res) {
-    controllers.getAccessToken(req, res, service);
-  }
-  // code gen
+  // static async seattlePost(req, res) {
+  //   controllers.seattlePost(req, res, service);
+  // }
 }
 
 module.exports = ThreadController;
