+++ modules/bot/router.js
@@ -1,9 +1,8 @@
 const controller = require("./controller");
 const router = require("express").Router();
-const ProfileModel = include("domain/models/profile");
 
 
-router.post("/github/post", controller.githubPost);
+router.post("/github/post", controller.gitHubPost);
 router.post("/seattle/post", controller.seattlePost);
 
 // code gen
